requirements.txt
absl-py==2.1.0
accelerate==0.33.0
addict==2.4.0
aiohttp==3.9.1
aiosignal==1.3.1
alembic==1.13.0
annotated-types==0.7.0
asttokens 
async-timeout==4.0.3
attrs==23.1.0
Automat==22.10.0
bioc==2.1
blinker==1.7.0
buildtools==1.0.6
causal-conv1d 
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
comm 
constantly==23.10.4
contourpy==1.2.0
cycler==0.12.1
Cython 
databricks-cli==0.18.0
datasets==2.15.0
debugpy 
decorator 
deepspeed==0.15.0
dill==0.3.7
docker==6.1.3
docopt==0.6.2
easydict==1.13
einops==0.7.0
entrypoints 
et-xmlfile==1.1.0
exceptiongroup 
executing 
filelock==3.13.1
Flask==3.0.0
fonttools==4.46.0
frozenlist==1.4.0
fsspec==2023.10.0
furl==2.1.3
fvcore==0.1.5.post20221221
gitdb==4.0.11
GitPython==3.1.40
grad-cam==1.5.3
greenlet==3.0.2
grpcio==1.64.1
gunicorn==21.2.0
hjson==3.1.0
huggingface-hub==0.24.6
hyperlink==21.0.0
idna==3.6
importlib-metadata==7.0.0
incremental==22.10.0
intervaltree==3.1.0
iopath==0.1.10
ipykernel 
ipython 
itsdangerous==2.1.2
jedi 
jieba==0.42.1
Jinja2==3.1.2
joblib==1.3.2
jsonlines==4.0.0
jsonpickle==3.3.0
jupyter-client 
jupyter_core 
kiwisolver==1.4.5
lightning==2.4.0
lightning-utilities==0.11.6
loguru==0.7.2
lxml==5.3.0
Mako==1.3.0
mamba-ssm 
Markdown==3.5.1
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline 
mlflow==2.9.1
mmcv==1.3.8
mmsegmentation==0.14.1
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
munkres==1.1.4
nest_asyncio 
networkx==3.2.1
ninja==1.11.1.1
numpy==1.26.2
nvidia-ml-py==12.560.30
oauthlib==3.2.2
opencv-python==4.8.1.78
openpyxl==3.1.2
orderedmultidict==1.0.1
packaging==23.2
pandas==2.1.3
parso 
peft==0.12.0
pexpect 
pickleshare 
Pillow==10.1.0
platformdirs==4.1.0
portalocker==2.8.2
prettytable==3.9.0
prompt_toolkit 
protobuf==4.25.1
psutil 
ptyprocess 
pure_eval 
py-cpuinfo==9.0.0
pyarrow==14.0.1
pyarrow-hotfix==0.6
pydantic==2.8.2
pydantic_core==2.20.1
Pygments 
PyJWT==2.8.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-hostlist==1.23.0
pytoolconfig==1.3.1
pytorch-lightning==2.4.0
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq 
querystring-parser==1.2.4
qwen-vl-utils==0.0.4
redo==2.0.4
regex==2023.10.3
requests==2.31.0
rope==1.13.0
safetensors==0.4.1
scikit-learn==1.3.2
scipy==1.11.4
sentencepiece==0.2.0
simplejson==3.19.2
six 
smmap==5.0.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.23
sqlparse==0.4.4
stack-data 
sympy==1.12
tabulate==0.9.0
tensorboard==2.17.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
termcolor==2.4.0
threadpoolctl==3.2.0
tiktoken==0.7.0
timm==0.4.12
tokenizers==0.19.1
tomli==2.0.1
torch==2.1.1+cu118
torchaudio==2.1.1+cu118
torchcam==0.4.0
torchmetrics==1.4.1
torchvision==0.16.1+cu118
tornado 
tqdm==4.66.1
traitlets 
transformers==4.45.0.dev0
transformers-stream-generator==0.0.5
triton==2.1.0
ttach==0.0.3
Twisted==23.10.0
typing_extensions 
tzdata==2023.3
urllib3==2.1.0
wcwidth==0.2.12
websocket-client==1.7.0
Werkzeug==3.0.1
xxhash==3.4.1
yacs==0.1.8
yapf==0.40.2
yarl==1.9.4
zipp==3.17.0
zope.interface==6.2